+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add "Use implicit message sending" refactoring for replacing explicit message sending calls with a `uses/2` directive
 * Add completions for the first argument of the `logtalk_load/1-2` predicates based on the standard libraries
 * Add completions for the first argument of the `logtalk_load_context/2` predicate
+* Add completions for the arguments of the `current_logtalk_flag/2` and `set_logtalk_flag/2` predicates
 * Remove the "Logtalk: Toggle Debugging" button from the editor title bar (subsumed by the "Run" menu commands)
 * Remove the "Logtalk:" prefix from all context menu items
 * Add actionable status bar indicators for profiling and CodeLens on/off state